move plugin types
